Argonaut API - moved resource tests to appropriate package
