+++ argonaut-api/src/test/java/gov/va/api/health/argonaut/api/resources/AllergyIntoleranceTest.java
@@ -1,4 +1,4 @@
-package gov.va.api.health.argonaut.api;
+package gov.va.api.health.argonaut.api.resources;
 
 import static gov.va.api.health.argonaut.api.RoundTrip.assertRoundTrip;
