+++ argonaut-api/src/test/java/gov/va/api/health/argonaut/api/resources/DiagnosticReportTest.java
@@ -1,4 +1,4 @@
-package gov.va.api.health.argonaut.api.resources.diagnosticreport;
+package gov.va.api.health.argonaut.api.resources;
 
 import static gov.va.api.health.argonaut.api.RoundTrip.assertRoundTrip;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -6,7 +6,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import gov.va.api.health.argonaut.api.ZeroOrOneVerifier;
 import gov.va.api.health.argonaut.api.bundle.AbstractBundle;
 import gov.va.api.health.argonaut.api.bundle.BundleLink;
-import gov.va.api.health.argonaut.api.resources.DiagnosticReport;
 import gov.va.api.health.argonaut.api.samples.SampleDiagnosticReports;
 import java.util.Collections;
 import java.util.Set;
