+++ argonaut-api/src/test/java/gov/va/api/health/argonaut/api/resources/MedicationTest.java
@@ -1,10 +1,9 @@
-package gov.va.api.health.argonaut.api;
+package gov.va.api.health.argonaut.api.resources;
 
 import static gov.va.api.health.argonaut.api.RoundTrip.assertRoundTrip;
 
 import gov.va.api.health.argonaut.api.bundle.AbstractBundle;
 import gov.va.api.health.argonaut.api.bundle.BundleLink;
-import gov.va.api.health.argonaut.api.resources.Medication;
 import gov.va.api.health.argonaut.api.samples.SampleMedications;
 import java.util.Collections;
 import lombok.extern.slf4j.Slf4j;
